fix sleb128 reading

+++ spec.hpp
@@ -62,9 +62,9 @@ static inline int64_t read_sleb128(uint8_t *&iter) {
 template <typename T, uint8_t bits = sizeof(T) * 8>
 static inline T safe_read_sleb128(uint8_t *&iter) {
     assert(bits / 8 <= sizeof(T));
-    uint64_t result = read_sleb128(iter);
-    assert(result <= (1ULL << (bits - 1)) - 1);
-    assert(result >= -(1ULL << (bits - 1)));
+    int64_t result = read_sleb128(iter);
+    assert(result <= static_cast<int64_t>((1LL << (bits - 1)) - 1));
+    assert(result >= static_cast<int64_t>(-(1LL << (bits - 1))));
     return static_cast<T>(result);
 }
